requirements.txt

fastapi
requests
uvicorn[standard]
keyvalue_sqlite
pytz
types-python-dateutil
python-dateutil
